+++ app/resources/translations/sr.ts
@@ -1,0 +1,819 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE TS>
+<TS version="2.1" language="sr_RS">
+<context>
+    <name>AboutDialog</name>
+    <message>
+        <location filename="../../about_ui/aboutdialog.ui" line="14"/>
+        <source>About mpz</source>
+        <translation>О mpz</translation>
+    </message>
+    <message>
+        <location filename="../../about_ui/aboutdialog.ui" line="29"/>
+        <location filename="../../about_ui/aboutdialog.cpp" line="68"/>
+        <source>Changelog</source>
+        <translation>Цхангелог</translation>
+    </message>
+    <message>
+        <location filename="../../about_ui/aboutdialog.ui" line="36"/>
+        <source>About Qt</source>
+        <translation>О Qt</translation>
+    </message>
+    <message>
+        <location filename="../../about_ui/aboutdialog.ui" line="43"/>
+        <source>Contact with developer</source>
+        <translation>Контактирајте програмера</translation>
+    </message>
+    <message>
+        <location filename="../../about_ui/aboutdialog.cpp" line="33"/>
+        <source>Using opensource libraries</source>
+        <translation>Коришћење библиотека отвореног кода</translation>
+    </message>
+    <message>
+        <location filename="../../about_ui/aboutdialog.cpp" line="37"/>
+        <source>Config files path: </source>
+        <translation>Путања конфигурационих датотека: </translation>
+    </message>
+</context>
+<context>
+    <name>BusySpinner</name>
+    <message>
+        <location filename="../../busyspinner.cpp" line="18"/>
+        <source>Background operation running</source>
+        <translation>Радња у позадини је покренута</translation>
+    </message>
+</context>
+<context>
+    <name>DirectorySettings</name>
+    <message>
+        <location filename="../../directory_ui/directorysettings.ui" line="14"/>
+        <source>Library folders</source>
+        <translation>Фасцикле библиотеке</translation>
+    </message>
+    <message>
+        <location filename="../../directory_ui/directorysettings.ui" line="27"/>
+        <source>Add</source>
+        <translation>Додати</translation>
+    </message>
+    <message>
+        <location filename="../../directory_ui/directorysettings.ui" line="34"/>
+        <source>Remove</source>
+        <translation>Уклони</translation>
+    </message>
+</context>
+<context>
+    <name>DirectoryUi::DirectoryContextMenu</name>
+    <message>
+        <location filename="../../directory_ui/directorycontextmenu.cpp" line="30"/>
+        <source>Clear filter</source>
+        <translation>Обриши филтер</translation>
+    </message>
+    <message>
+        <location filename="../../directory_ui/directorycontextmenu.cpp" line="32"/>
+        <source>Create new playlist</source>
+        <translation>Креирајте нову листу песама</translation>
+    </message>
+    <message>
+        <location filename="../../directory_ui/directorycontextmenu.cpp" line="34"/>
+        <source>Append to current playlist</source>
+        <translation>Додај тренутној плејлисти</translation>
+    </message>
+    <message>
+        <location filename="../../directory_ui/directorycontextmenu.cpp" line="35"/>
+        <source>Open in file manager</source>
+        <translation>Отворите у менаџеру датотека</translation>
+    </message>
+</context>
+<context>
+    <name>DirectoryUi::SortMenu</name>
+    <message>
+        <location filename="../../directory_ui/directorysortmenu.cpp" line="17"/>
+        <source>Default (Name)</source>
+        <translation>Подразумевано (име)</translation>
+    </message>
+</context>
+<context>
+    <name>FeedbackForm</name>
+    <message>
+        <location filename="../../feedback_ui/feedbackform.ui" line="14"/>
+        <source>Feedback form</source>
+        <translation>Образац за повратне информације</translation>
+    </message>
+    <message>
+        <location filename="../../feedback_ui/feedbackform.ui" line="20"/>
+        <source>Found a bug? Need a feature? Tell developer what you think about mpz</source>
+        <translation>Пронашли сте грешку? Треба вам функција? Реците програмеру шта мислите о mpz</translation>
+    </message>
+    <message>
+        <location filename="../../feedback_ui/feedbackform.ui" line="30"/>
+        <source>(optional )Your name / contact if you&apos;d like to hear back</source>
+        <translation>(опционо) Ваше име/контакт ако желите да се јавите</translation>
+    </message>
+    <message>
+        <location filename="../../feedback_ui/feedbackform.ui" line="46"/>
+        <source>Include anonymous system info</source>
+        <translation>Укључите анонимне системске информације</translation>
+    </message>
+    <message>
+        <location filename="../../feedback_ui/feedbackform.ui" line="56"/>
+        <source>This will help, but you can opt-out</source>
+        <translation>Ово ће помоћи, али можете да одустанете</translation>
+    </message>
+    <message>
+        <location filename="../../feedback_ui/feedbackform.ui" line="66"/>
+        <source>Send</source>
+        <translation>Пошаљи</translation>
+    </message>
+    <message>
+        <location filename="../../feedback_ui/feedbackform.cpp" line="35"/>
+        <source>Thanks for you feedback! (click again to close)</source>
+        <translation>Хвала на повратним информацијама! (кликните поново да затворите)</translation>
+    </message>
+    <message>
+        <location filename="../../feedback_ui/feedbackform.cpp" line="68"/>
+        <source>Error occured, please try again</source>
+        <translation>Дошло је до грешке, покушајте поново</translation>
+    </message>
+</context>
+<context>
+    <name>MainMenu</name>
+    <message>
+        <location filename="../../mainmenu.cpp" line="17"/>
+        <source>Tray icon</source>
+        <translation>Икона у траци</translation>
+    </message>
+    <message>
+        <location filename="../../mainmenu.cpp" line="20"/>
+        <source>Minimize to tray</source>
+        <translation>Минимизирајте у послужавник</translation>
+    </message>
+    <message>
+        <location filename="../../mainmenu.cpp" line="25"/>
+        <source>Playback log</source>
+        <translation>Дневник репродукције</translation>
+    </message>
+    <message>
+        <location filename="../../mainmenu.cpp" line="26"/>
+        <source>About mpz</source>
+        <translation>О mpz</translation>
+    </message>
+    <message>
+        <location filename="../../mainmenu.cpp" line="27"/>
+        <source>Quit</source>
+        <translation>Одустати</translation>
+    </message>
+    <message>
+        <location filename="../../mainmenu.cpp" line="28"/>
+        <source>Got feedback?</source>
+        <translation>Имате повратне информације?</translation>
+    </message>
+    <message>
+        <location filename="../../mainmenu.cpp" line="29"/>
+        <source>Keyboard shortcuts</source>
+        <translation>Пречице на тастатури</translation>
+    </message>
+    <message>
+        <location filename="../../mainmenu.cpp" line="30"/>
+        <source>Save settings</source>
+        <translation>Сачувај подешавања</translation>
+    </message>
+</context>
+<context>
+    <name>MainWindow</name>
+    <message>
+        <location filename="../../mainwindow.ui" line="14"/>
+        <source>MainWindow</source>
+        <translation>МаинВиндов</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="23"/>
+        <source>Stop</source>
+        <translation>Зауставити</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="30"/>
+        <source>Pause</source>
+        <translation>Пауза</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="37"/>
+        <source>Play</source>
+        <translation>Играј</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="44"/>
+        <source>Previous track</source>
+        <translation>Претходна нумера</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="51"/>
+        <source>Next track</source>
+        <translation>Следећа нумера</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="73"/>
+        <source>Track duration</source>
+        <translation>Трајање нумере</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="95"/>
+        <source>Background operation running</source>
+        <translation>Радња у позадини је покренута</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="102"/>
+        <source>Sort</source>
+        <translation>Врста</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="112"/>
+        <source>Volume</source>
+        <translation>Гласноћа</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="125"/>
+        <source>Main menu</source>
+        <translation>Главни мени</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="146"/>
+        <source>Configure library folders</source>
+        <translation>Конфигуришите фасцикле библиотеке</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="149"/>
+        <source>⚙</source>
+        <translation>⚙</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="156"/>
+        <source>Select library folder</source>
+        <translation>Изаберите фасциклу библиотеке</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="172"/>
+        <source>Filter library</source>
+        <translation>Библиотека филтера</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="186"/>
+        <source>Filter playlists</source>
+        <translation>Филтрирајте листе песама</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="209"/>
+        <source>Filter tracks</source>
+        <translation>Филтрирајте трагове</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="216"/>
+        <source>Playback follows selection</source>
+        <translation>Репродукција следи избор</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="219"/>
+        <source>Playback follows cursor</source>
+        <translation>Репродукција прати курсор</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="226"/>
+        <source>Order:</source>
+        <translation>Наручите:</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="233"/>
+        <source>global</source>
+        <translation>глобално</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="240"/>
+        <source>Default playback order</source>
+        <translation>Подразумевани редослед репродукције</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="247"/>
+        <source>playlist</source>
+        <translation>плаилист</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.ui" line="254"/>
+        <source>Override playback for current playlist</source>
+        <translation>Замени репродукцију за тренутну листу песама</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.cpp" line="152"/>
+        <location filename="../../mainwindow.cpp" line="174"/>
+        <source>sequential</source>
+        <translation>секвенцијално</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.cpp" line="153"/>
+        <location filename="../../mainwindow.cpp" line="173"/>
+        <source>random</source>
+        <translation>насумично</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.cpp" line="172"/>
+        <source>(use global)</source>
+        <translation>(користите глобално)</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.cpp" line="302"/>
+        <location filename="../../mainwindow.cpp" line="306"/>
+        <source>Nothing selected</source>
+        <translation>Ништа није изабрано</translation>
+    </message>
+    <message>
+        <location filename="../../mainwindow.cpp" line="308"/>
+        <source>Selection total duration</source>
+        <translation>Укупно трајање избора</translation>
+    </message>
+</context>
+<context>
+    <name>PlaybackLogDialog</name>
+    <message>
+        <location filename="../../playback_log_ui/playbacklogdialog.ui" line="14"/>
+        <source>Playback log</source>
+        <translation>Дневник репродукције</translation>
+    </message>
+    <message>
+        <location filename="../../playback_log_ui/playbacklogdialog.ui" line="25"/>
+        <location filename="../../playback_log_ui/playbacklogdialog.cpp" line="93"/>
+        <source>Save as CSV</source>
+        <translation>Сачувај као CSV</translation>
+    </message>
+    <message>
+        <location filename="../../playback_log_ui/playbacklogdialog.cpp" line="47"/>
+        <source>Total time played</source>
+        <translation>Укупно одиграно време</translation>
+    </message>
+    <message>
+        <location filename="../../playback_log_ui/playbacklogdialog.cpp" line="51"/>
+        <source>This session time played</source>
+        <translation>Ово време сесије је одиграно</translation>
+    </message>
+    <message>
+        <location filename="../../playback_log_ui/playbacklogdialog.cpp" line="69"/>
+        <source>Copy</source>
+        <translation>Копирај</translation>
+    </message>
+    <message>
+        <location filename="../../playback_log_ui/playbacklogdialog.cpp" line="73"/>
+        <source>Jump to</source>
+        <translation>Скочи на</translation>
+    </message>
+    <message>
+        <location filename="../../playback_log_ui/playbacklogdialog.cpp" line="78"/>
+        <source>Search on web</source>
+        <translation>Претражите на вебу</translation>
+    </message>
+</context>
+<context>
+    <name>PlaylistUi::PlaylistContextMenu</name>
+    <message>
+        <location filename="../../playlist_ui/playlistcontextmenu.cpp" line="18"/>
+        <source>Remove</source>
+        <translation>Уклони</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/playlistcontextmenu.cpp" line="22"/>
+        <source>Show in file manager</source>
+        <translation>Прикажи у менаџеру датотека</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/playlistcontextmenu.cpp" line="26"/>
+        <source>Copy name</source>
+        <translation>Копирај име</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/playlistcontextmenu.cpp" line="29"/>
+        <source>Clear filter</source>
+        <translation>Обриши филтер</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/playlistcontextmenu.cpp" line="33"/>
+        <source>Track info</source>
+        <translation>Информације о нумери</translation>
+    </message>
+</context>
+<context>
+    <name>PlaylistsUi::PlaylistsContextMenu</name>
+    <message>
+        <location filename="../../playlists_ui/playlistscontextmenu.cpp" line="28"/>
+        <source>Remove</source>
+        <translation>Уклони</translation>
+    </message>
+    <message>
+        <location filename="../../playlists_ui/playlistscontextmenu.cpp" line="29"/>
+        <source>Rename</source>
+        <translation>Преименуј</translation>
+    </message>
+    <message>
+        <location filename="../../playlists_ui/playlistscontextmenu.cpp" line="30"/>
+        <location filename="../../playlists_ui/playlistscontextmenu.cpp" line="93"/>
+        <source>Save as m3u</source>
+        <translation>Сачувај као m3u</translation>
+    </message>
+    <message>
+        <location filename="../../playlists_ui/playlistscontextmenu.cpp" line="31"/>
+        <source>Reload from filesystem</source>
+        <translation>Поново учитај из система датотека</translation>
+    </message>
+    <message>
+        <location filename="../../playlists_ui/playlistscontextmenu.cpp" line="32"/>
+        <source>Play</source>
+        <translation>Играј</translation>
+    </message>
+    <message>
+        <location filename="../../playlists_ui/playlistscontextmenu.cpp" line="52"/>
+        <source>Clear filter</source>
+        <translation>Обриши филтер</translation>
+    </message>
+    <message>
+        <location filename="../../playlists_ui/playlistscontextmenu.cpp" line="83"/>
+        <source>Rename playlist</source>
+        <translation>Преименујте плејлисту</translation>
+    </message>
+</context>
+<context>
+    <name>Shortcuts</name>
+    <message>
+        <location filename="../../shortcuts.cpp" line="33"/>
+        <source>Play</source>
+        <translation>Играј</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="34"/>
+        <source>Stop</source>
+        <translation>Зауставити</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="35"/>
+        <source>Pause</source>
+        <translation>Пауза</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="36"/>
+        <source>Next</source>
+        <translation>Следећа</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="37"/>
+        <source>Previous</source>
+        <translation>Претходна</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="38"/>
+        <source>Focus on library</source>
+        <translation>Фокусирајте се на библиотеку</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="39"/>
+        <source>Focus on playlists</source>
+        <translation>Фокусирајте се на листе песама</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="40"/>
+        <source>Focus on playlist</source>
+        <translation>Фокусирајте се на листу за репродукцију</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="41"/>
+        <source>Focus on library filter</source>
+        <translation>Фокусирајте се на филтер библиотеке</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="42"/>
+        <source>Focus on playlists filter</source>
+        <translation>Фокусирајте се на филтер плејлиста</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="43"/>
+        <source>Focus on playlist filter</source>
+        <translation>Фокусирајте се на филтер листе песама</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="44"/>
+        <source>Open main menu</source>
+        <translation>Отворите главни мени</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="45"/>
+        <source>Open playback log</source>
+        <translation>Отворите дневник репродукције</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="46"/>
+        <source>Open sort menu</source>
+        <translation>Отворите мени за сортирање</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="47"/>
+        <source>Open shortcuts dialog</source>
+        <translation>Отворите дијалог пречица</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="48"/>
+        <source>Jump to playing track</source>
+        <translation>Пређите на нумеру за репродукцију</translation>
+    </message>
+    <message>
+        <location filename="../../shortcuts.cpp" line="49"/>
+        <source>Quit</source>
+        <translation>Одустати</translation>
+    </message>
+</context>
+<context>
+    <name>ShortcutsDialog</name>
+    <message>
+        <location filename="../../shortcuts_ui/shortcutsdialog.ui" line="14"/>
+        <source>Keyboard shortcuts</source>
+        <translation>Пречице на тастатури</translation>
+    </message>
+</context>
+<context>
+    <name>SortUi::SortMenu</name>
+    <message>
+        <location filename="../../sort_ui/sortmenu.cpp" line="35"/>
+        <source>Default</source>
+        <translation>Уобичајено</translation>
+    </message>
+    <message>
+        <location filename="../../sort_ui/sortmenu.cpp" line="53"/>
+        <source>Edit presets</source>
+        <translation>Уредите унапред подешене поставке</translation>
+    </message>
+</context>
+<context>
+    <name>SortingPresets</name>
+    <message>
+        <location filename="../../sort_ui/sortingpresets.ui" line="14"/>
+        <source>Sorting presets</source>
+        <translation>Сортирање унапред подешених поставки</translation>
+    </message>
+    <message>
+        <location filename="../../sort_ui/sortingpresets.ui" line="22"/>
+        <source>Help</source>
+        <translation>Помоћ</translation>
+    </message>
+    <message>
+        <location filename="../../sort_ui/sortingpresets.ui" line="29"/>
+        <source>New preset</source>
+        <translation>Ново унапред подешено</translation>
+    </message>
+    <message>
+        <location filename="../../sort_ui/sortingpresets.ui" line="36"/>
+        <source>Test</source>
+        <translation>Тест</translation>
+    </message>
+    <message>
+        <location filename="../../sort_ui/sortingpresets.ui" line="52"/>
+        <source>Add</source>
+        <translation>Додати</translation>
+    </message>
+    <message>
+        <location filename="../../sort_ui/sortingpresets.ui" line="66"/>
+        <source>Rename</source>
+        <translation>Преименуј</translation>
+    </message>
+    <message>
+        <location filename="../../sort_ui/sortingpresets.ui" line="73"/>
+        <source>Remove</source>
+        <translation>Уклони</translation>
+    </message>
+</context>
+<context>
+    <name>SortingPresetsDialog</name>
+    <message>
+        <location filename="../../sort_ui/sortingpresetsdialog.cpp" line="14"/>
+        <source>Default</source>
+        <translation>Уобичајено</translation>
+    </message>
+    <message>
+        <location filename="../../sort_ui/sortingpresetsdialog.cpp" line="52"/>
+        <source>Rename sorting preset</source>
+        <translation>Преименујте унапред подешено сортирање</translation>
+    </message>
+    <message>
+        <location filename="../../sort_ui/sortingpresetsdialog.cpp" line="103"/>
+        <source>Sorting presets description</source>
+        <translation>Опис сортирања унапред подешених поставки</translation>
+    </message>
+</context>
+<context>
+    <name>StatusBarLabel</name>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="18"/>
+        <location filename="../../statusbarlabel.cpp" line="38"/>
+        <source>Stopped</source>
+        <translation>Заустављен</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="24"/>
+        <source>Playing</source>
+        <translation>Играње</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="31"/>
+        <source>Paused</source>
+        <translation>Паузирано</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="56"/>
+        <source>Copy</source>
+        <translation>Копирај</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="60"/>
+        <source>Show playback log</source>
+        <translation>Прикажи дневник репродукције</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="62"/>
+        <source>Jump to playing track</source>
+        <translation>Пређите на нумеру за репродукцију</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="64"/>
+        <source>Search on web</source>
+        <translation>Претражите на вебу</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="81"/>
+        <source>stream buffer</source>
+        <translation>бафер тока</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="88"/>
+        <source>KB</source>
+        <translation>КБ</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="88"/>
+        <source>MB</source>
+        <translation>МБ</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="88"/>
+        <source>GB</source>
+        <translation>ГБ</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="88"/>
+        <source>TB</source>
+        <translation>ТБ</translation>
+    </message>
+    <message>
+        <location filename="../../statusbarlabel.cpp" line="91"/>
+        <source>bytes</source>
+        <translation>бајтова</translation>
+    </message>
+</context>
+<context>
+    <name>TrackInfoDialog</name>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.ui" line="14"/>
+        <source>Track info</source>
+        <translation>Информације о нумери</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.ui" line="41"/>
+        <source>Show in file manager</source>
+        <translation>Прикажи у менаџеру датотека</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="47"/>
+        <source>Artist</source>
+        <translation>Уметник</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="50"/>
+        <source>Album</source>
+        <translation>Албум</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="53"/>
+        <source>Title</source>
+        <translation>Наслов</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="56"/>
+        <source>Year</source>
+        <translation>Година</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="59"/>
+        <source>Track number</source>
+        <translation>Број нумере</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="60"/>
+        <source>Duration</source>
+        <translation>Трајање</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="62"/>
+        <source>Format</source>
+        <translation>Формат</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="63"/>
+        <source>Bitrate</source>
+        <translation>Битрате</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="64"/>
+        <source>Sample rate</source>
+        <translation>Стопа узорковања</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="65"/>
+        <source>Channels</source>
+        <translation>Канали</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="67"/>
+        <source>Stream url</source>
+        <translation>URL стрима</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="70"/>
+        <source>File path</source>
+        <translation>Путања до датотеке</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="73"/>
+        <source>CUE start at</source>
+        <translation>CUE почиње у</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="95"/>
+        <source>Copy</source>
+        <translation>Копирај</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="99"/>
+        <source>Search on web</source>
+        <translation>Претражите на вебу</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="127"/>
+        <source>Copy to clipboard</source>
+        <translation>Копирај у међуспремник</translation>
+    </message>
+    <message>
+        <location filename="../../playlist_ui/trackinfodialog.cpp" line="135"/>
+        <source>Open in external viewer</source>
+        <translation>Отвори у спољном прегледнику</translation>
+    </message>
+</context>
+<context>
+    <name>TrayIcon</name>
+    <message>
+        <location filename="../../trayicon.cpp" line="9"/>
+        <location filename="../../trayicon.cpp" line="75"/>
+        <source>Stopped</source>
+        <translation>Заустављен</translation>
+    </message>
+    <message>
+        <location filename="../../trayicon.cpp" line="12"/>
+        <source>Quit</source>
+        <translation>Одустати</translation>
+    </message>
+    <message>
+        <location filename="../../trayicon.cpp" line="14"/>
+        <source>Play</source>
+        <translation>Играј</translation>
+    </message>
+    <message>
+        <location filename="../../trayicon.cpp" line="16"/>
+        <source>Pause</source>
+        <translation>Пауза</translation>
+    </message>
+    <message>
+        <location filename="../../trayicon.cpp" line="18"/>
+        <source>Stop</source>
+        <translation>Зауставити</translation>
+    </message>
+    <message>
+        <location filename="../../trayicon.cpp" line="20"/>
+        <source>Next</source>
+        <translation>Следећа</translation>
+    </message>
+    <message>
+        <location filename="../../trayicon.cpp" line="22"/>
+        <source>Previous</source>
+        <translation>Претходна</translation>
+    </message>
+    <message>
+        <location filename="../../trayicon.cpp" line="65"/>
+        <source>Playing</source>
+        <translation>Играње</translation>
+    </message>
+    <message>
+        <location filename="../../trayicon.cpp" line="85"/>
+        <source>Paused</source>
+        <translation>Паузирано</translation>
+    </message>
+</context>
+</TS>